+++ samples/qyl.sample/Program.cs
@@ -2,7 +2,7 @@ using Qyl.Sample;
 
 var builder = WebApplication.CreateSlimBuilder(args);
 
-builder.Services.AddQylApi(AppJsonSerializerContext.Default);
+builder.AddQylApi(AppJsonSerializerContext.Default);
 builder.Services.AddSingleton<TodoStore>();
 
 var app = builder.Build();